backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==13.1

# For testing
pytest==8.3.3

llama-cpp-python==0.2.90
requests==2.32.3

